+++ python/scoundrel_python/__init__.py
@@ -1,3 +1,12 @@
+from .scoundrel_json import ScoundrelTypeHandler, dumps as scoundrel_json_dumps, loads as scoundrel_json_loads, register_scoundrel_type
 from .web_socket_server import ScoundrelPythonServer, WebSocketClient, main
 
-__all__ = ["ScoundrelPythonServer", "WebSocketClient", "main"]
+__all__ = [
+  "ScoundrelPythonServer",
+  "WebSocketClient",
+  "main",
+  "ScoundrelTypeHandler",
+  "register_scoundrel_type",
+  "scoundrel_json_dumps",
+  "scoundrel_json_loads"
+]